RTAS: adjust column width for "Argument Call Buffer" tables
Each RTAS command has a corresponding table documenting input/output
values. The 1st column is either "In" or "Out", the 2nd is the
name of a parameter, and the 3rd is the description. Rather than
using equal widths for each, use 1*/2*/4* proportions to use the
table space a bit more efficiently. This reduces current document
size from 211 pages to 207.

Signed-off-by: Michael Roth <mdroth@linux.vnet.ibm.com>